Add support for aarch64 musl (#569)
Closes #87 

Does not include the `+static` variant as I was having a hard time
getting the bzip2 build passing. I'll open a new issue to track that.

Related https://github.com/astral-sh/python-build-standalone/pull/484

+++ cpython-unix/build-zstd.sh
@@ -15,16 +15,20 @@ tar -xf zstd-${ZSTD_VERSION}.tar.gz
 pushd cpython-source-deps-zstd-${ZSTD_VERSION}/lib
 
 if [ "${CC}" = "musl-clang" ]; then
-    # In order to build the library with SSE2, BMI, and AVX2 intrinstics, we need musl-clang to find
+    # In order to build the library with intrinsics, we need musl-clang to find
     # headers that provide access to the intrinsics, as they are not provided by musl. These are
     # part of the include files that are part of clang. But musl-clang eliminates them from the
     # default include path. So copy them into place.
-    for h in ${TOOLS_PATH}/${TOOLCHAIN}/lib/clang/*/include/*intrin.h ${TOOLS_PATH}/${TOOLCHAIN}/lib/clang/*/include/{__wmmintrin_aes.h,__wmmintrin_pclmul.h,emmintrin.h,immintrin.h,mm_malloc.h}; do
+    for h in ${TOOLS_PATH}/${TOOLCHAIN}/lib/clang/*/include/*intrin.h ${TOOLS_PATH}/${TOOLCHAIN}/lib/clang/*/include/{__wmmintrin_aes.h,__wmmintrin_pclmul.h,emmintrin.h,immintrin.h,mm_malloc.h,arm_neon.h,arm_neon_sve_bridge.h,arm_bf16.h,arm_fp16.h,arm_acle.h,arm_vector_types.h}; do
         filename=$(basename "$h")
-        if [ -e "${TOOLS_PATH}/host/include/${filename}" ]; then
-            echo "warning: ${filename} already exists"
+        if [ -f "$h" ]; then
+            if [ -e "${TOOLS_PATH}/host/include/${filename}" ]; then
+                echo "warning: ${filename} already exists"
+            fi
+            cp "$h" ${TOOLS_PATH}/host/include/
+        else
+            echo "warning: ${filename} not found (skipping)"
         fi
-        cp "$h" ${TOOLS_PATH}/host/include/
     done
     EXTRA_TARGET_CFLAGS="${EXTRA_TARGET_CFLAGS} -I${TOOLS_PATH}/host/include/"
 
@@ -33,7 +37,7 @@ if [ "${CC}" = "musl-clang" ]; then
     # `qsort_r` is actually available so we patch it to include a check for glibc.
     patch -p1 <<EOF
 diff --git a/dictBuilder/cover.c b/dictBuilder/cover.c
-index 5e6e8bc..6ca72a1 100644
+index 2ef33c7..078e2ee 100644
 --- a/dictBuilder/cover.c
 +++ b/dictBuilder/cover.c
 @@ -241,7 +241,7 @@ typedef struct {
@@ -45,6 +49,24 @@ index 5e6e8bc..6ca72a1 100644
  /* C90 only offers qsort() that needs a global context. */
  static COVER_ctx_t *g_coverCtx = NULL;
  #endif
+@@ -290,7 +290,7 @@ static int COVER_cmp8(COVER_ctx_t *ctx, const void *lp, const void *rp) {
+  */
+ #if (defined(_WIN32) && defined(_MSC_VER)) || defined(__APPLE__)
+ static int WIN_CDECL COVER_strict_cmp(void* g_coverCtx, const void* lp, const void* rp) {
+-#elif defined(_GNU_SOURCE)
++#elif defined(_GNU_SOURCE) && defined(__GLIBC__)
+ static int COVER_strict_cmp(const void *lp, const void *rp, void *g_coverCtx) {
+ #else /* C90 fallback.*/
+ static int COVER_strict_cmp(const void *lp, const void *rp) {
+@@ -306,7 +306,7 @@ static int COVER_strict_cmp(const void *lp, const void *rp) {
+  */
+ #if (defined(_WIN32) && defined(_MSC_VER)) || defined(__APPLE__)
+ static int WIN_CDECL COVER_strict_cmp8(void* g_coverCtx, const void* lp, const void* rp) {
+-#elif defined(_GNU_SOURCE)
++#elif defined(_GNU_SOURCE) && defined(__GLIBC__)
+ static int COVER_strict_cmp8(const void *lp, const void *rp, void *g_coverCtx) {
+ #else /* C90 fallback.*/
+ static int COVER_strict_cmp8(const void *lp, const void *rp) {
 @@ -328,7 +328,7 @@ static void stableSort(COVER_ctx_t *ctx) {
      qsort_r(ctx->suffix, ctx->suffixSize, sizeof(U32),
              ctx,
